backend/requirements.txt
fastapi==0.95.1
pydantic==1.10.7
sqlalchemy==2.0.10
psycopg2-binary==2.9.6
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
uvicorn==0.22.0
email-validator==2.0.0.post2
